Refactor charts
